requirements.txt
requests
beautifulsoup4
python-dotenv
pandas
tqdm
pytest
matplotlib
pyarrow
